Fix type import error
