+++ web-common/src/components/navigation/breadcrumbs/types.ts
@@ -1,0 +1,10 @@
+export type Param = string;
+
+export type PathOption = {
+  label: string;
+  depth?: number;
+  href?: string;
+  section?: string;
+};
+
+export type PathOptions = Map<Param, PathOption>;
